requirements.txt
fastapi==0.95.1
gunicorn==20.1.0
Flask==2.3.1
Flask-Cors==3.0.10
openai==0.27.4
langchain==0.0.142
pinecone-client==2.2.1
Werkzeug==2.3.2
python-dotenv==1.0.0
aiohttp==3.8.4
black==23.3.0
fastapi==0.95.1
huggingface-hub==0.14.0
numpy==1.23.5
nltk==3.8.1
pandas==1.5.3
Pillow==9.5.0
pdfplumber==0.9.0
pytesseract==0.3.10
PyYAML==5.4.1
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
SQLAlchemy==1.4.47
tiktoken==0.3.3
torch==2.0.0
torchvision==0.15.1
transformers==4.27.4
tqdm==4.65.0
uvicorn==0.21.1
unstructured==0.6.1
python-multipart==0.0.6
